Creacion of Function obtenerReparto adn modifications in index.js file
