+++ entregas/Grupo04/index.js
@@ -7,7 +7,7 @@ const app = express();
 const bodyParser = require('body-parser');
 app.use(bodyParser.json());
 
-const { leerTrailerflix, obtenerTrailerPorId, obtenerTitulo } = require('./src/trailerflix.controller');
+const { leerTrailerflix, obtenerTrailerPorId, obtenerTitulo, obtenerReparto } = require('./src/trailerflix.controller');
 
 const PORT = process.env.PORT || 3000;
 
@@ -40,13 +40,7 @@ app.get('/api/reparto/:act', (req, res) => {
     let param = req.params['act'].trim().toLowerCase();
     console.log(param)
     if (param !== ""){
-        let result = DB.filter( item => 
-            item.reparto.some(reparto => 
-            reparto.toLowerCase().includes(param))).map(item => 
-            ({
-            titulo: item.titulo,
-            reparto: item.reparto
-        }));
+        let result = obtenerReparto(param, DB);
 
         result.length > 0 ? res.json(result) :
         res.status(404).json({ id: 'Error', descripcion: 'Ups!!! No hay coincidencias encontradas :(' })
